+++ test/js/util/mapbox.test.js
@@ -10,12 +10,12 @@ test("mapbox", function(t) {
 
     t.test('.normalizeSourceURL', function(t) {
         t.test('returns a v4 URL with access_token parameter', function(t) {
-            t.equal(mapbox.normalizeSourceURL('mapbox://user.map'), 'https://a.tiles.mapbox.com/v4/user.map.json?access_token=key&secure');
+            t.equal(mapbox.normalizeSourceURL('mapbox://user.map'), 'https://api.mapbox.com/v4/user.map.json?access_token=key&secure');
             t.end();
         });
 
         t.test('uses provided access token', function(t) {
-            t.equal(mapbox.normalizeSourceURL('mapbox://user.map', 'token'), 'https://a.tiles.mapbox.com/v4/user.map.json?access_token=token&secure');
+            t.equal(mapbox.normalizeSourceURL('mapbox://user.map', 'token'), 'https://api.mapbox.com/v4/user.map.json?access_token=token&secure');
             t.end();
         });
 
@@ -33,13 +33,6 @@ test("mapbox", function(t) {
             t.end();
         });
 
-        t.test('omits &secure and uses http when FORCE_HTTPS is false', function(t) {
-            config.FORCE_HTTPS = false;
-            t.equal(mapbox.normalizeSourceURL('mapbox://user.map'), 'http://a.tiles.mapbox.com/v4/user.map.json?access_token=key');
-            config.FORCE_HTTPS = true;
-            t.end();
-        });
-
         t.test('ignores non-mapbox:// scheme', function(t) {
             t.equal(mapbox.normalizeSourceURL('http://path'), 'http://path');
             t.end();
@@ -52,7 +45,7 @@ test("mapbox", function(t) {
         t.test('correctly transforms mapbox:// URLs', function(t) {
             t.equal(
                 mapbox.normalizeGlyphsURL('mapbox://fonts/boxmap/{fontstack}/{range}.pbf'),
-                'https://a.tiles.mapbox.com/fonts/v1/boxmap/{fontstack}/{range}.pbf?access_token=key'
+                'https://api.mapbox.com/fonts/v1/boxmap/{fontstack}/{range}.pbf?access_token=key'
             );
             t.end();
         });
@@ -67,7 +60,7 @@ test("mapbox", function(t) {
 
     t.test('.normalizeStyleURL', function(t) {
         t.test('returns an API URL with access_token parameter', function(t) {
-            t.equal(mapbox.normalizeStyleURL('mapbox://user.style'), 'https://a.tiles.mapbox.com/styles/v1/user/user.style?access_token=key');
+            t.equal(mapbox.normalizeStyleURL('mapbox://user.style'), 'https://api.mapbox.com/styles/v1/user/user.style?access_token=key');
             t.end();
         });